Clean up the connection to hueblobs when hueblobs quits
The GIOChannel emits a "HUP" event when hueblobs dies.  This patch
makes the viewier catch that, close up everything to do with the
hueblobs connection, and start retrying to open the connection to it.

Without this patch, if hueblobs quits whilst interacting with squidge,
then squidge is left in an infinite loop of processor use.