Don't read and discard a byte from the FIFO when data's ready
Whilst devleoping the camera view, before it was parsing the data
hueblobs was sending it, I had it set up to just read a single byte
from the fifo in order to avoid infinite loops.  Turns out that I
forgot to remove this single byte reading, resulting in the first blob
being drawn in the wrong place.  I didn't notice this because I was,
in general, testing with many many blobs visible in noisy
environments...